+++ d1-master/Project_Economic/src/main/java/com/example/project_economic/dto/HistoryCartDto.java
@@ -20,7 +20,7 @@ public class HistoryCartDto {
     private LocalDateTime BoughtAt;
 
     private Integer quantity;
-    private Double sumTotalInCartHistory(){
+    private Long sumTotalInCartHistory(){
         return this.quantity*this.productResponse.getCostPrice();
     }
